Now for a bit of yak shaving.
Overload Make_Point and Make_Vector with integer arguments, so I don't
have to type ".0" after each literal when it is a whole number.

Only works if all 3 arguments are whole numbers. Technically I could
overload the functions with each possible permutation of floats and
integers but I don't think it is happeing.