A bit of cleanup after trying to add a restrict.adc similar to the one in Stanislav Datskovskiy's libFFA.
Wrote Radatracer.Canvas.Clamp to comply with No_Implicit_Conditional.

Removed Radatracer.Canvas.Make_Canvas to comply with No_Secondary_Stack.
Apparently functions that return unconstrained objects rely on the heap.

Unfortunately I cannot use this because AUnit relies on tons and tons of features,
including implicit conditionals, secondary stack usage and many more. And even
if restrict.adc is only referenced in radatracer.gpr it also applies to any downstream
project that links with libradatracer, so the way it works is that it forces me to
remove pragmas that libradatracer complies with, which is kind of pointless in this project.