simplify design, I need to move on.
For now Sphere is the only "object" and it is a regular record. This
way I don't have to deal with OOP and pointerolade until there is an
actual hierarchy of classes.

Also switched back to Vectors, I think that for these use cases it'll
be way less clunky to use than unconstrained arrays.